+++ app/services/anoubis/redis_services/init.rb
@@ -1,4 +1,4 @@
-module Anoubis::RedisServices::Init
+class Anoubis::RedisServices::Init < Anoubis::ApplicationService
   # @!attribute default_host
   #   @return [String] default Redis database host (default: 127.0.0.1)
   attr_accessor :redis_host
@@ -11,8 +11,16 @@ module Anoubis::RedisServices::Init
   # @!attribute default_prefix
   #  @return [String] default redis prefix (default: '')
   attr_accessor :redis_prefix
+  # @!attribute key
+  #   @return [String] Redis key
+  attr_accessor :key
 
-  def setup
+  ##
+  # Initialize service
+  # @param [String] key Redis key (default: '')
+  def initialize(key = '')
+    @key = key
+
     begin
       @redis_host = Rails.configuration.redis_host
     rescue
@@ -53,5 +61,12 @@ module Anoubis::RedisServices::Init
   # @return [Redis] Redis database connection
   def redis
     Redis.new( host: redis_host, port: redis_port, db: redis_db )
+  end
+
+  ##
+  # Return redis key with prefix
+  # @return [String] redis key with prefix
+  def key
+    "#{redis_prefix}#{@key}"
   end
 end