+++ app/services/anoubis/redis_services/set.rb
@@ -1,17 +1,9 @@
-class Anoubis::RedisServices::Set < Anoubis::ApplicationService
-  include Anoubis::RedisServices::Init
-
-  def initialize(key)
-    setup
-
-    @key = "#{redis_prefix}#{key}"
-  end
-
+class Anoubis::RedisServices::Set < Anoubis::RedisServices::Init
   def call(data, time = nil)
     if (data.class == Array) || (data.class == Hash)
-      redis.set(@key, data.to_json, ex: time) == 'OK'
+      redis.set(key, data.to_json, ex: time) == 'OK'
     else
-      redis.set(@key, data, ex: time) == 'OK'
+      redis.set(key, data, ex: time) == 'OK'
     end
   end
 end